requirements.txt
tensorflow==2.18.0
numpy==1.26.4
pillow==10.4.0
scikit-learn==1.4.2
matplotlib==3.8.4
tqdm==4.66.4
streamlit==1.39.0

